+++ eyes.common.java/src/main/java/com/applitools/eyes/AccessibilityLevel.java
@@ -1,7 +1,6 @@
 package com.applitools.eyes;
 
 public enum AccessibilityLevel {
-    None,
     AA,
     AAA
 }